+++ lib/tilt/etanni.rb
@@ -5,7 +5,7 @@ module Tilt
   class EtanniTemplate < Template
     def prepare
       separator = data.hash.abs
-      chomp = "<<#{separator}.chomp!"
+      chomp = "<<#{separator}.chomp"
       start = "\n_out_ << #{chomp}\n"
       stop = "\n#{separator}\n"
       replacement = "#{stop}\\1#{start}"
@@ -13,7 +13,7 @@ module Tilt
       temp = @data.strip
       temp.gsub!(/<\?r\s+(.*?)\s+\?>/m, replacement)
 
-      @code = "_out_ = [<<#{separator}.chomp!]\n#{temp}#{stop}_out_.join"
+      @code = "_out_ = [<<#{separator}.chomp]\n#{temp}#{stop}_out_.join"
     end
 
     def precompiled_template(locals)