+++ lib/tilt/wikicloth.rb
@@ -2,6 +2,8 @@
 require_relative 'template'
 require 'wikicloth'
 
+warn 'tilt/wikicloth is deprecated, as wikicloth requires modifying string literals', uplevel: 1
+
 # WikiCloth implementation. See: https://github.com/nricciar/wikicloth
 Tilt::WikiClothTemplate = Tilt::StaticTemplate.subclass do
   parser = @options.delete(:parser) || WikiCloth::Parser