+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 * Support commonmarker 1.0+ API (unasuke) (#10)
+* Make etanni template work with frozen string literals (jeremyevans)
 
 ## 2.3.0 (2023-09-14)